Replace resource ID switches in sample app
Convert switch statements on R.id values to if/else chains so the sample app builds with non-final Android resource IDs used by newer AGP versions.

+++ OpenStreetMapViewer/src/main/java/org/osmdroid/PreferenceActivity.java
@@ -281,35 +281,21 @@ public class PreferenceActivity extends AppCompatActivity implements View.OnClic
     @Override
     public void onClick(View v) {
 
-        switch (v.getId()) {
-            case R.id.buttonManualCacheEntry: {
-                showManualEntry(textViewCacheDirectory);
-            }
-            break;
-            case R.id.buttonSetCache: {
-                showPickCacheFromList(textViewCacheDirectory, "tiles" + File.separator);
-            }
-            break;
-            case R.id.buttonPurgeCache: {
-                purgeCache();
-            }
-            break;
-            case R.id.buttonReset: {
-                resetSettings(this);
-                abortSave = true;
-                finish();
-            }
-            break;
-            case R.id.buttonManualBaseEntry: {
-                showManualEntry(textViewBaseDirectory);
-            }
-            break;
-            case R.id.buttonSetBase: {
-                showPickCacheFromList(textViewBaseDirectory, "");
-            }
-            break;
-
-
+        int id = v.getId();
+        if (id == R.id.buttonManualCacheEntry) {
+            showManualEntry(textViewCacheDirectory);
+        } else if (id == R.id.buttonSetCache) {
+            showPickCacheFromList(textViewCacheDirectory, "tiles" + File.separator);
+        } else if (id == R.id.buttonPurgeCache) {
+            purgeCache();
+        } else if (id == R.id.buttonReset) {
+            resetSettings(this);
+            abortSave = true;
+            finish();
+        } else if (id == R.id.buttonManualBaseEntry) {
+            showManualEntry(textViewBaseDirectory);
+        } else if (id == R.id.buttonSetBase) {
+            showPickCacheFromList(textViewBaseDirectory, "");
         }
     }
